mv AvgOverTime to avg_over_time.go
